Rename propertiesSeed.ts to propertySeed.ts
